path_planning.cpp: update the logic to find the next target to clean.
When finding on the direction toward Y+, we only search the '10' target,
until it allows shift to Y-, it also search the '01' target, since it
beyonds the current robot's Y position.

Same as above, when finding on the direction toward Y+, we only search
the '01' target, until it allows shift to Y-, it also search the '01'
target, since it beyonds the current robot's Y position.