Change to find the shorest path from robot to target.
Originally, when finding the shortest path from robot to target, we use
the A-start shortest path from the current robot position, until we
reach the target position, but now, we switch to from the target
position back to the robot position. The advantage for this is that we
can have better control of we direction that the robot should go.