Fix potential forced the robot in trapped error.
The error is due to set the current robot position with in
path_escape_trapped(), since now we are using SLAM, the position of the
robot keeps jumping, in path_next(), if a target can't be reach, then it
will try to call path_escape_trapped() to verify the robot is trapped or
not, in this case, if we set the current position by using
path_set_current_pos() inside path_escape_trapped(), it will cause the
consistance error of the position. So, better set position before
entering path_escape_trapped().