Change go home logic in cleanning mode.
Now robot will save the position where it sees the charger stub in a list call "Home_Point". When it
finish cleanning or remote controled to go home, it will go to the first position in the list to see if it
can still see the charger stub. If succeed, it will try go to the charger stub. If failed, it will try next
position. The last position is the start point of robot(0, 0).